Move AppLinkButton component to elements dir
