+++ components/AppWishListButton.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import AppLinkButton from '~/components/AppLinkButton'
+import AppLinkButton from '~/components/elements/AppLinkButton'
 
 const { WISH_LIST_URL } = process.env.config
