+++ pages/index.vue
@@ -84,7 +84,7 @@ import AppBooks from '~/components/AppBooks'
 import AppGifteeButton from '~/components/AppGifteeButton'
 import AppHeader1 from '~/components/AppHeader1'
 import AppHeader2 from '~/components/AppHeader2'
-import AppLinkButton from '~/components/AppLinkButton'
+import AppLinkButton from '~/components/elements/AppLinkButton'
 import AppPayPayButton from '~/components/AppPayPayButton'
 import AppProfileIcon from '~/components/AppProfileIcon'
 import AppSkillButton from '~/components/AppSkillButton'
